[ISSUE #3501]🚀Implement GeneralHAService and controller mode for unified HA architecture✨
